requirements.txt
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
cython==0.29.2
falcon==1.4.1
gunicorn==19.9.0
httpie==1.0.2
ipython==7.2.0
ipython-genutils==0.2.0
more-itertools==5.0.0
msgpack-python==0.5.6
pluggy==0.8.1
py==1.7.0
pylint==2.2.2
pytest==4.1.1
ujson==1.35